+++ realtime-reporter/pkg/handlers/watch.go
@@ -8,13 +8,13 @@ import (
 func WatchHandler(resourceType string) cache.ResourceEventHandlerFuncs {
 
 	var handler cache.ResourceEventHandlerFuncs
-	handler.AddFunc = func(obj interface{}) {
+	handler.AddFunc = func(obj any) {
 		logrus.WithField("resourceType", resourceType).WithField("obj", obj).Info("add event")
 	}
-	handler.UpdateFunc = func(old, new interface{}) {
+	handler.UpdateFunc = func(old, new any) {
 		logrus.WithField("resourceType", resourceType).WithField("old", old).WithField("new", new).Info("update event")
 	}
-	handler.DeleteFunc = func(obj interface{}) {
+	handler.DeleteFunc = func(obj any) {
 		logrus.WithField("resourceType", resourceType).WithField("obj", obj).Info("delete event")
 	}
 	return handler