+++ plugins/ci/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 6.2.8
+* Bumped to Go 1.26
+* Bump library dependencies
+* Bump indirect library dependencies
+
 ## 6.2.7
 * Bump helm to 4.1.1
 * Bump library k8s.io/api, k8s.io/apimachinery, k8s.io/client-go to v0.35.1